% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{sample of random data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
data
}
\description{
@format A data frame that contains three columns of meaningless data
 \describe{
   \item{a}{A list of scenarios}
   \item{b}{just numbers}
   \item{c}{more numbers}
 }
 @source my head
}
\keyword{datasets}
