% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{(numeric)}
}
\value{
The temperature provided in Fahrenheit converted to Celsius
}
\description{
Convert Fahrenheit to Celsius
}
\examples{
fahrenheit_to_celsius(fahrenheit=70)
}
