% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_to_cm.R
\name{in_to_cm}
\alias{in_to_cm}
\title{Convert Inches to Centimeters}
\usage{
in_to_cm(inches)
}
\arguments{
\item{inches}{(numeric) A positive number}
}
\value{
The length provided in inches converted to centimeters
}
\description{
The purpose of this function is to convert inches into centimeters.
}
\examples{
in_to_cm(inches=1)
}
