% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_function}
\alias{test_function}
\title{This is a test function}
\usage{
test_function(input)
}
\arguments{
\item{input}{put an input (numeric)}
}
\description{
this is a test function to practice my package building
}
\examples{
test_function(10)
}
